Add support for default Merger in Config::factory()
In a config array, one can pass an array of MergerInterface objects
in the 'mergers' key, but if the merger is not an object, the factory()
will instantiate a new ValueMerger() with this value as constructor parameter.
This means that the parameter has to be a valid MergePolicy identifier.

Example:

```php
return [

    'mergers' => [
	// will use a new ValueMerger(MergePolicy::COMBINE) to handle the packages.loaded directive
        'packages.loaded' => MergePolicy::COMBINE
    ],
    'directives' => [
	// your configuration directives here
    ]
];
```